Fix missing categories in header by moving data fetching to server side
