+++ components/Header.tsx
@@ -1,142 +1,18 @@
-'use client'
-
 import Link from 'next/link'
-import { useState, useEffect } from 'react'
 import { getCategories } from '@/lib/cosmic'
 import { Category } from '@/types'
 import SearchInput from './SearchInput'
+import ClientHeader from './ClientHeader'
 
-export default function Header() {
-  const [categories, setCategories] = useState<Category[]>([])
-  const [isMobileMenuOpen, setIsMobileMenuOpen] = useState(false)
-  const [isLoading, setIsLoading] = useState(true)
-
-  useEffect(() => {
-    async function fetchCategories() {
-      try {
-        const categoriesData = await getCategories()
-        setCategories(categoriesData as Category[])
-      } catch (error) {
-        console.error('Error fetching categories:', error)
-      } finally {
-        setIsLoading(false)
-      }
-    }
-    
-    fetchCategories()
-  }, [])
-
-  const toggleMobileMenu = () => {
-    setIsMobileMenuOpen(!isMobileMenuOpen)
+export default async function Header() {
+  let categories: Category[] = []
+  
+  try {
+    categories = await getCategories() as Category[]
+  } catch (error) {
+    console.error('Error fetching categories:', error)
+    // Continue with empty categories array
   }
 
-  const closeMobileMenu = () => {
-    setIsMobileMenuOpen(false)
-  }
-
-  if (isLoading) {
-    return (
-      <header className="sticky top-0 z-50 bg-white/95 dark:bg-gray-900/95 border-b border-gray-200 dark:border-gray-800 backdrop-blur-sm transition-colors">
-        <div className="container">
-          <div className="flex items-center justify-between h-16">
-            <Link href="/" className="text-xl font-bold text-gray-900 dark:text-white">
-              Blog Website
-            </Link>
-            <div className="flex items-center space-x-4">
-              <SearchInput />
-              <div className="md:hidden">
-                <button className="text-gray-600 dark:text-gray-300">
-                  <svg className="w-6 h-6" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-                    <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M4 6h16M4 12h16M4 18h16" />
-                  </svg>
-                </button>
-              </div>
-            </div>
-          </div>
-        </div>
-      </header>
-    )
-  }
-
-  return (
-    <header className="sticky top-0 z-50 bg-white/95 dark:bg-gray-900/95 border-b border-gray-200 dark:border-gray-800 backdrop-blur-sm transition-colors">
-      <div className="container">
-        <div className="flex items-center justify-between h-16">
-          <Link href="/" className="text-xl font-bold text-gray-900 dark:text-white">
-            Blog Website
-          </Link>
-          
-          {/* Desktop Navigation */}
-          <nav className="hidden md:flex items-center space-x-8">
-            <Link 
-              href="/" 
-              className="text-gray-600 dark:text-gray-300 hover:text-gray-900 dark:hover:text-white transition-colors"
-            >
-              Home
-            </Link>
-            
-            {categories.slice(0, 3).map((category) => (
-              <Link
-                key={category.id}
-                href={`/categories/${category.slug}`}
-                className="text-gray-600 dark:text-gray-300 hover:text-gray-900 dark:hover:text-white transition-colors"
-              >
-                {category.metadata?.name || category.title}
-              </Link>
-            ))}
-          </nav>
-
-          {/* Search Input and Mobile Menu Button */}
-          <div className="flex items-center space-x-4">
-            <SearchInput />
-            
-            {/* Mobile menu button */}
-            <div className="md:hidden">
-              <button 
-                onClick={toggleMobileMenu}
-                className="text-gray-600 dark:text-gray-300 hover:text-gray-900 dark:hover:text-white transition-colors"
-                aria-label="Toggle mobile menu"
-              >
-                {isMobileMenuOpen ? (
-                  <svg className="w-6 h-6" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-                    <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M6 18L18 6M6 6l12 12" />
-                  </svg>
-                ) : (
-                  <svg className="w-6 h-6" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-                    <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M4 6h16M4 12h16M4 18h16" />
-                  </svg>
-                )}
-              </button>
-            </div>
-          </div>
-        </div>
-
-        {/* Mobile Navigation Menu */}
-        {isMobileMenuOpen && (
-          <div className="md:hidden border-t border-gray-200 dark:border-gray-800 bg-white dark:bg-gray-900">
-            <nav className="py-4 space-y-2">
-              <Link 
-                href="/" 
-                className="block px-4 py-2 text-gray-600 dark:text-gray-300 hover:text-gray-900 dark:hover:text-white hover:bg-gray-50 dark:hover:bg-gray-800 transition-colors"
-                onClick={closeMobileMenu}
-              >
-                Home
-              </Link>
-              
-              {categories.map((category) => (
-                <Link
-                  key={category.id}
-                  href={`/categories/${category.slug}`}
-                  className="block px-4 py-2 text-gray-600 dark:text-gray-300 hover:text-gray-900 dark:hover:text-white hover:bg-gray-50 dark:hover:bg-gray-800 transition-colors"
-                  onClick={closeMobileMenu}
-                >
-                  {category.metadata?.name || category.title}
-                </Link>
-              ))}
-            </nav>
-          </div>
-        )}
-      </div>
-    </header>
-  )
+  return <ClientHeader categories={categories} />
 }
